api/requirements.txt
fastapi
uvicorn
pika
psycopg2-binary
redis
